fix(push): review firebase token removal

+++ src/framework/modules/auth/service.ts
@@ -495,11 +495,10 @@ export class FcmService {
         this._removeTokenFromDeleteQueue(token);
         console.debug('FcmService - unregisterFCMToken - OK - ', token);
       } else {
-        console.debug('FcmService - unregisterFCMToken - NO TOKEN - ');
+        console.debug('FcmService - unregisterFCMToken - NO TOKEN');
       }
     } catch (err) {
-      console.error('FcmService - unregisterFCMToken - ERROR - ', token);
-      console.error((err as Error).message);
+      console.error('FcmService - unregisterFCMToken - ERROR - ', (err as Error).message);
       //unregistering fcm token should not crash the login process
       if (!Connection.isOnline) {
         //when no connection => get it from property
@@ -514,20 +513,22 @@ export class FcmService {
       if (!token) {
         token = await messaging().getToken();
       }
-      const req = OAuth2RessourceOwnerPasswordClient.signRequestWithToken(
-        OAuth2RessourceOwnerPasswordClient.convertTokenToOldObjectSyntax(account.tokens),
-        `${this.platform.url}/timeline/pushNotif/fcmToken?fcmToken=${token}`,
-        {
-          method: 'delete',
-        },
-      );
-
-      await fetch(req);
-      this._removeTokenFromDeleteQueue(token);
-      console.debug('FcmService - unregisterFCMTokenWithAccount - OK - ', token);
+      if (token) {
+        const req = OAuth2RessourceOwnerPasswordClient.signRequestWithToken(
+          OAuth2RessourceOwnerPasswordClient.convertTokenToOldObjectSyntax(account.tokens),
+          `${this.platform.url}/timeline/pushNotif/fcmToken?fcmToken=${token}`,
+          {
+            method: 'delete',
+          },
+        );
+        await fetch(req);
+        this._removeTokenFromDeleteQueue(token);
+        console.debug('FcmService - unregisterFCMTokenWithAccount - OK - ', token);
+      } else {
+        console.debug('FcmService - unregisterFCMToken - NO TOKEN');
+      }
     } catch (err) {
-      console.error('FcmService - unregisterFCMTokenWithAccount - ERROR - ', token);
-      console.error((err as Error).message);
+      console.error('FcmService - unregisterFCMTokenWithAccount - ERROR - ', (err as Error).message);
       //unregistering fcm token should not crash the login process
       if (!Connection.isOnline) {
         //when no connection => get it from property
